(auto)adding cRAP contaminants to sequence database
-Renamed module namespace to "proteomics"
-Implemented test
