+++ src/fastadl/config.vsh.yaml
@@ -1,0 +1,39 @@
+functionality:
+  name: fastadl
+  namespace: proteomics
+  description: Downloads a Fasta sequence database for a given taxonomy ID
+  usage: fastadl --taxid 9606 --output db/
+  authors:
+    - name: Kenneth Verheggen
+      email: kenneth@data-intuitive.com
+      roles: [ maintainer ]
+      props: { github: KennethVerheggen }
+  argument_groups:
+    - name: Inputs
+      arguments:
+        - name: "--taxid"
+          type: string
+          required: true
+          description: The taxonomyID to obtain a fasta for
+          example: "9606"
+        - name: "--output"
+          type: file
+          required: true
+          description: An output directory for the fasta files
+          example: "db/"
+        - name: "--include_contaminants"
+          type: boolean
+          default: true
+          description: Include a collection of common contaminants from thegpmdb (highly advised)
+  resources:
+    - type: python_script
+      path: script.py
+  tests:
+    - type: bash_script
+      path: test.sh
+platforms:
+  - type: docker
+    image: python:3.8
+    setup:
+    - type: python
+      pypi: requests
