--- src/uniprot/config.vsh.yaml
@@ -1,30 +1,0 @@
-functionality:
-  name: fastadl
-  namespace: uniprot
-  description: Downloads a Fasta sequence database for a given taxonomy ID
-  usage: fastadl --taxid 9606 --output db/
-  authors:
-    - name: Kenneth Verheggen
-      email: kenneth@data-intuitive.com
-      roles: [ maintainer ]
-      props: { github: KennethVerheggen }
-  arguments:
-    - name: "--taxid"
-      type: string
-      required: true
-      description: The taxonomyID to obtain a fasta for
-      example: "9606"
-    - name: "--output"
-      type: file
-      required: true
-      description: An output directory for the fasta files
-      example: "db/"
-  resources:
-    - type: python_script
-      path: script.py
-platforms:
-  - type: docker
-    image: python:3.8
-    setup:
-    - type: python
-      pypi: requests
